+++ nana.Game/Screens/NanaScreensTypes.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace nanaGame.Screens
+{
+    enum NanaScreensTypes
+    {
+        MainMenu,
+        Game,
+        ResultScreen
+    }
+}