+++ src/game/interfaces/audio.interface.ts
@@ -1,6 +1,0 @@
-interface IAudio {
-  scene: Phaser.Scene;
-  beatmapMusic: string;
-}
-
-export { IAudio };