+++ src/game/interfaces/sceneTransition.interface.ts
@@ -1,4 +1,0 @@
-export interface ISceneTransition {
-  scene: Phaser.Scene;
-  isShow: boolean;
-}