+++ src/game/interfaces/topBar.interface.ts
@@ -1,4 +1,0 @@
-export interface ITopBar {
-  scene: Phaser.Scene;
-  onBackClick: () => void;
-}