+++ src/game/interfaces/userConfig.interface.ts
@@ -1,7 +1,0 @@
-export interface IIserConfig {
-  showNoteAccuracy: boolean;
-  showPerfectHit: boolean;
-  hitPosition: number;
-  musicVolume: number;
-  hitsoundVolume: number;
-}