+++ src/game/redux/userConfig/getters/index.ts
@@ -1,7 +1,0 @@
-import store from '../../store';
-
-const getUserConfig = () => {
-  return store.getState().userConfig;
-};
-
-export { getUserConfig };